+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [1.79.0] - 2020-11-13
+-   Add requestType in get of HttpService.
+-   Add RESTService::getPartialBufferFromServer method to retrieve a peace of data from server.
+-   Fix in presenceEventHandler the offline presence of a contact in the roster.
+-   Add FileServerService::getBlobFromUrlWithOptimizationObserver method to retrieve a file and store it in a path. An Observer is returned to follow the download progress.
+-   Add FileStorageService::downloadFileInPath method to retrieve a file and store it in a path. An Observer is returned to follow the download progress.
+-   Fix FileStorageService::downloadFile for big files. So the returned Object.buffer is modified to an Array of data.
+
 ## [1.79.0-dotnet.0] - 2020-11-04
 -   Update the AlertsService documentation